lab-2/requirements.txt
strawberry-graphql~=0.269.0
fastapi~=0.115.12
SQLAlchemy~=2.0.41
pymongo~=4.13.0
motor~=3.7.1